requirements.txt
Mastodon.py>=1.8.1
requests>=2.31.0
python-dotenv>=1.0.0
thefuzz>=0.22.1
python-Levenshtein>=0.25.0
